+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Adds Sentry generator. ([@mausamp][])
 * Adds Dotenv generator. ([@mausamp][])
 * Adds Honeybadger generator. ([@mausamp][])
+* Adds Rails ERD generator. ([@mausamp][])
 
 ## 0.13.0 (March 26th, 2024)
 * Adds Letter Opener generator. ([@coolprobn][])